+++ frontend/src/components/@shared/GoBackButton/GoBackButton.styles.ts
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+const Button = styled.button`
+  position: absolute;
+  top: 1rem;
+  left: 0.5rem;
+  background-color: transparent;
+  border: 0;
+  cursor: pointer;
+`;
+
+export { Button };
